Refactor transaction addition flow by creating AddTransactionPage component
