+++ src/app/assetitems/[assetItemId]/transactions/add/page.tsx
@@ -1,49 +1,11 @@
-'use client';
-import SidebarTriggerWithBreadcrumb from '@/components/SidebarTriggerWithBreadcrumb';
-import { Card } from '@/components/ui/card';
-import AddTransactionForm from '@/features/transactions/components/AddTransactionForm';
-import { withAssetItemPortfolios } from '@/features/portfolio/hoc/withAssetItemPortfolios';
-import withAssetItems from '@/features/assetItems/hoc/withAssetItems';
-import withCurrency from '@/components/hoc/withCurrency';
-import { AssetItemPortfolio } from '@/types';
+import AddTransactionPage from '@/features/transactions/components/AddTransactionPage';
 
-export default function Page({ params }: { params: { assetItemId: string } }) {
-	const assetItemId = params.assetItemId;
-
-	const WithAddTransactionFormWrapper = withAssetItems(
-		withCurrency(withAssetItemPortfolios(AddTransactionFormWrapper))
-	);
-
-	return (
-		<WithAddTransactionFormWrapper assetItemIds={[assetItemId]} latest={true} />
-	);
-}
-
-function AddTransactionFormWrapper({
-	portfolios,
+export default async function Page({
+	params,
 }: {
-	portfolios: AssetItemPortfolio[];
+	params: Promise<{ assetItemId: string }>;
 }) {
-	const assetItem = portfolios[0];
+	const { assetItemId } = await params;
 
-	return (
-		<>
-			<title>Add Transaction</title>
-			<SidebarTriggerWithBreadcrumb
-				breadcrumbs={[
-					{ title: 'Asset Items', href: '/assetitems' },
-					{ title: assetItem.name, href: `/assetitems/${assetItem.id}` },
-					{
-						title: 'Add Transaction',
-						href: `/assetitems/${assetItem.id}/transactions/add`,
-					},
-				]}
-			/>
-			<div className='container mx-auto p-2 h-full'>
-				<Card className='p-8 max-w-screen-sm mx-auto'>
-					<AddTransactionForm assetItem={assetItem} />
-				</Card>
-			</div>
-		</>
-	);
+	return <AddTransactionPage assetItemId={assetItemId} />;
 }
